+++ src/main/java/com/example/egoverment/Employee.java
@@ -6,7 +6,7 @@ public class Employee extends Human{
     private String startingDate;
 
 
-    public Employee(String name, int id){
+    public Employee(String name, long id){
         super(name,id);
         startingDate=LocalDate.now().toString();
     }
